Move some helper methods into helpers module
